requirements.txt
streamlit
numpy
sentence-transformers
faiss-cpu
PyMuPDF
pillow
bcrypt
plotly
scikit-learn
python-docx
